Splits tests
