+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.tests.js
@@ -121,10 +121,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.deepEqual(dataGrid.option('columns'), ['field1', { dataField: 'field2' }], 'columns option is not changed');
     });
 
-    QUnit.test('formatValue for grouped column with calculateGroupValue', function(assert) {
-        assert.strictEqual(gridCore.formatValue('2012', { format: 'shortDate' }), '2012');
-    });
-
     QUnit.test('commonColumnOptions', function(assert) {
         const dataGrid = createDataGrid({});
         assert.deepEqual(dataGrid.option('commonColumnSettings'), {
@@ -387,27 +383,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         }
     });
 
-    QUnit.test('Cells in group row should not have \'dx-col-fixed\' class (T852898)', function(assert) {
-    // arrange
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: {
-                store: [
-                    { id: 1, value: 'value 1' }
-                ]
-            },
-            columns: ['id', {
-                dataField: 'value',
-                fixed: true,
-                groupIndex: 0
-            }]
-        }).dxDataGrid('instance');
-
-        // assert
-        assert.notOk($(dataGrid.getCellElement(0, 0)).hasClass('dx-col-fixed'), 'dx-col-fixed');
-        assert.notOk($(dataGrid.getCellElement(0, 1)).hasClass('dx-col-fixed'), 'dx-col-fixed');
-    });
-
     QUnit.test('Rows with \'dx-row-alt\' should not have \'dx-col-fixed\' class on cells (T852898)', function(assert) {
     // arrange
         const dataGrid = $('#dataGrid').dxDataGrid({
@@ -440,120 +415,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         // assert
         assert.strictEqual(dataGrid.getView('rowsView').option('noDataText'), noDataText, 'valid noDataText in rowsView options');
     });
-
-    QUnit.test('Change row expand state on row click', function(assert) {
-        let isRowClicked = false;
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2'],
-            loadingTimeout: undefined,
-            onRowPrepared: function(args) {
-                assert.equal(typeUtils.isRenderer(args.rowElement), !!config().useJQuery, 'rowElement is correct');
-                if(args.rowType === 'group') {
-                    if(isRowClicked) {
-                        assert.ok(!args.component.isRowExpanded(args.key), 'after click on group row it\'s closed');
-                    } else {
-                        assert.ok(args.component.isRowExpanded(args.key), 'group row is expanded');
-                    }
-                }
-            },
-            grouping: {
-                expandMode: 'rowClick'
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '1', field2: '4' }],
-                group: 'field1'
-            }
-        }).dxDataGrid('instance');
-
-        isRowClicked = true;
-        $(dataGrid.$element())
-            .find('.dx-datagrid-rowsview tr')
-            .eq(0)
-            .trigger('dxclick');
-
-    });
-
-    QUnit.test('DataGrid - CustomStore.load should contain the \'select\' parameter after grouping operations (T817511)', function(assert) {
-    // arrange
-        const arrayStore = new ArrayStore({
-            key: 'id',
-            data: [
-                { id: 0, text: 'Text', countryId: 0, cityId: 0 },
-                { id: 1, text: 'Text', countryId: 0, cityId: 1 },
-                { id: 2, text: 'Text', countryId: 1, cityId: 0 },
-                { id: 3, text: 'Text', countryId: 1, cityId: 1 },
-            ]
-        });
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: {
-                key: 'id',
-                select: ['id', 'text', 'countryId', 'cityId'],
-                load: function(options) {
-                    const d = $.Deferred();
-
-                    // assert
-                    assert.notEqual(options.select, undefined, 'options.select is defined');
-
-                    setTimeout(function() {
-                        const result = {};
-                        arrayStore.load(options).done(function(data) {
-                            result.data = data;
-
-                            if(options.group) {
-                                data.forEach(item => {
-                                    item.count = item.items.length;
-                                    item.items = null;
-                                });
-                            }
-                        });
-                        if(options.requireGroupCount) {
-                            arrayStore.load({ filter: options.filter, group: options.group }).done(function(groupedData) {
-                                result.groupCount = groupedData.length;
-                            });
-                        }
-                        if(options.requireTotalCount) {
-                            arrayStore.totalCount(options).done(function(totalCount) {
-                                result.totalCount = totalCount;
-                            });
-                        }
-
-                        d.resolve(result);
-                    });
-
-                    return d;
-                },
-                group: [{
-                    selector: 'countryId',
-                    isExpanded: false
-                }, {
-                    selector: 'cityId',
-                    isExpanded: false
-                }]
-            },
-            remoteOperations: {
-                groupPaging: true
-            },
-            groupPanel: {
-                visible: true
-            },
-            grouping: {
-                autoExpandAll: false
-            }
-        }).dxDataGrid('instance');
-        this.clock.tick();
-
-        // act
-        dataGrid.expandRow([0]);
-        this.clock.tick();
-
-        dataGrid.expandRow([0, 0]);
-        this.clock.tick();
-
-        dataGrid.columnOption('cityId', 'groupIndex', undefined);
-        this.clock.tick();
-    });
-
 
     // T843705
     QUnit.test('DataSource should be reset after changing remoteOperations', function(assert) {
@@ -598,58 +459,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         }, 'loadOptions');
     });
 
-    // T756152
-    QUnit.test('CellTemplate should not be rendered on the group row when summary is enabled', function(assert) {
-    // arrange
-        const loadResult = [{
-            ID: 1,
-            OrderNumber: 777,
-            TotalValue: 12175,
-            BoolValue: true,
-            CustomerName: 'Test1'
-        }, {
-            ID: 4,
-            OrderNumber: 777,
-            TotalValue: 16550,
-            BoolValue: false,
-            CustomerName: 'Test1'
-        }];
-
-        $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: loadResult,
-            keyExpr: 'ID',
-            showBorders: true,
-            groupPanel: {
-                visible: true
-            },
-            columns: [
-                {
-                    dataField: 'CustomerName',
-                    groupIndex: 0
-                }, 'OrderNumber', 'TotalValue', {
-                    dataField: 'BoolValue',
-                    cellTemplate: function(container, options) {
-                        $('<div>').dxCheckBox({
-                            value: options.data.BoolValue,
-                            disabled: true
-                        }).appendTo(container);
-                    }
-                }
-            ],
-            summary: {
-                groupItems: [{
-                    column: 'TotalValue',
-                    summaryType: 'sum',
-                    alignByColumn: true
-                }]
-            }
-        });
-
-        // assert
-        assert.equal($('.dx-group-row').eq(0).find('.dx-checkbox').length, 0, 'group row does not contain checkbox');
-    });
-
     QUnit.test('cellClick/cellHoverChanged handler should be executed when define via \'on\' method', function(assert) {
         let cellClickCount = 0;
         let cellHoverChangedCount = 0;
@@ -683,326 +492,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
         assert.equal(cellClickCount, 1, 'Cell click is called once');
         assert.equal(cellHoverChangedCount, 2, 'Cell hover state changes 2 times');
-    });
-
-    QUnit.test('Default context menu shown when click on header panel items', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2'],
-            loadingTimeout: undefined,
-            editing: {
-                allowAdding: true
-            },
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '1', field2: '4' }]
-            }
-        }).dxDataGrid('instance');
-        const e = $.Event('dxcontextmenu');
-
-        $(dataGrid.$element())
-            .find('.dx-datagrid-addrow-button')
-            .trigger(e);
-
-        assert.notOk(e.isDefaultPrevented(), 'default behavior should not be prevented');
-    });
-
-    QUnit.test('Default context menu shown when click on command column', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2'],
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'row',
-                allowUpdating: true,
-                allowDeleting: true
-            },
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '1', field2: '4' }]
-            }
-        }).dxDataGrid('instance');
-        const e = $.Event('dxcontextmenu');
-
-        $(dataGrid.$element())
-            .find('.dx-header-row .dx-command-edit')
-            .first()
-            .trigger(e);
-
-        assert.notOk(e.isDefaultPrevented(), 'default behavior should not be prevented');
-    });
-
-    QUnit.test('Default context menu shown when click on rows', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2'],
-            loadingTimeout: undefined,
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '1', field2: '4' }]
-            }
-        }).dxDataGrid('instance');
-        const e = $.Event('dxcontextmenu');
-
-        $(dataGrid.$element())
-            .find('.dx-datagrid-rowsview .dx-row')
-            .first()
-            .trigger(e);
-
-        assert.notOk(e.isDefaultPrevented(), 'default behavior should not be prevented');
-    });
-
-    QUnit.test('Check grouping context menu operability', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2'],
-            loadingTimeout: undefined,
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '1', field2: '4' }]
-            }
-        }).dxDataGrid('instance');
-
-        $(dataGrid.$element())
-            .find('.dx-header-row td')
-            .eq(1)
-            .trigger('dxcontextmenu');
-
-        $('.dx-datagrid .dx-menu-item') // click on "group by this"
-            .eq(3)
-            .trigger('dxclick');
-
-        this.clock.tick(300);
-
-        assert.deepEqual(dataGrid.getController('data')._dataSource.group(), [{ selector: 'field2', desc: false, isExpanded: true }], 'datasource grouping is up to date');
-        assert.equal(dataGrid.columnOption('field2', 'groupIndex'), 0, 'Group by field2');
-
-        $(dataGrid.$element())
-            .find('.dx-header-row td')
-            .eq(1)
-            .trigger('dxcontextmenu');
-
-        $('.dx-datagrid .dx-menu-item') // click on "clear grouping"
-            .eq(4)
-            .trigger('dxclick');
-
-        this.clock.tick(300);
-
-        assert.equal(dataGrid.columnOption('field2', 'groupIndex'), undefined, 'field2 has no groupIndex');
-    });
-
-    QUnit.test('Show contextMenu for hidden adaptive columns', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            grouping: {
-                contextMenuEnabled: true
-            },
-            columnHidingEnabled: true,
-            width: 200,
-            dataSource: [
-                { field1: '1', field2: '2', field3: '3', field4: '4' },
-                { field1: '1', field2: '4', field3: '3', field4: '5' }
-            ]
-        }).dxDataGrid('instance');
-
-        $('.dx-datagrid .dx-datagrid-adaptive-more')
-            .eq(0)
-            .trigger('dxclick');
-
-        $('.dx-datagrid .dx-adaptive-detail-row')
-            .find('.dx-field-item-label-text')
-            .eq(0)
-            .trigger('dxcontextmenu');
-
-        const items = $('.dx-datagrid .dx-menu-item');
-        assert.equal(items.length, 5, 'context menu is generated');
-
-        items.eq(3).trigger('dxclick');
-
-        assert.deepEqual(dataGrid.getController('data')._dataSource.group(), [{ selector: 'field3', desc: false, isExpanded: true }], 'datasource grouping is up to date');
-        assert.equal(dataGrid.columnOption('field3', 'groupIndex'), 0, 'Group by field3');
-    });
-
-    QUnit.test('Group panel should set correct \'max-width\' after clear grouping', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: {
-                store: [
-                    { field1: '1', field2: '2', field3: '3', field4: '4', field5: '5' },
-                    { field1: '11', field2: '22', field3: '33', field4: '44', field5: '55' }]
-            },
-            width: 460,
-            groupPanel: {
-                emptyPanelText: 'Long long long long long long long long long long long text',
-                visible: true
-            },
-            editing: { allowAdding: true, mode: 'batch' },
-            columnChooser: {
-                enabled: true
-            }
-        }).dxDataGrid('instance');
-        const $dataGrid = $(dataGrid.element());
-
-        this.clock.tick();
-        assert.equal($dataGrid.find('.dx-toolbar-item-invisible').length, 4, '4 toolbar items are hidden, group panel has a long message');
-
-        dataGrid.columnOption('field2', 'groupIndex', 0);
-        this.clock.tick();
-
-        assert.equal($dataGrid.find('.dx-toolbar-item-invisible').length, 0, 'all toolbar items are visible, group panel has a group with short name');
-
-        dataGrid.clearGrouping();
-        this.clock.tick();
-        assert.equal($dataGrid.find('.dx-toolbar-item-invisible').length, 4, '4 toolbar items are hidden after clear grouping');
-    });
-
-    QUnit.test('Check grouping context menu operability (ungroup one column)', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', { dataField: 'field2', groupIndex: 1, showWhenGrouped: true }, { dataField: 'field3', groupIndex: 0, showWhenGrouped: true }],
-            sorting: {
-                mode: 'none'
-            },
-            loadingTimeout: undefined,
-            allowGrouping: true,
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2', field3: '34' }, { field1: '1', field2: '4', field3: '8' }]
-            }
-        }).dxDataGrid('instance');
-
-        $(dataGrid.$element())
-            .find('.dx-header-row td')
-            .eq(3)
-            .trigger('dxcontextmenu');
-
-        $('.dx-datagrid .dx-menu-item') // click on "Ungroup this"
-            .eq(1)
-            .trigger('dxclick');
-
-        this.clock.tick(300);
-
-        assert.equal(dataGrid.columnOption('field2', 'groupIndex'), undefined, 'field2 has no groupIndex');
-        assert.equal(dataGrid.columnOption('field3', 'groupIndex'), 0, 'field3 save its groupIndex');
-    });
-
-    QUnit.test('Ungroup one column via group row context menu', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', { dataField: 'field2', groupIndex: 1, showWhenGrouped: true }, { dataField: 'field3', groupIndex: 0, showWhenGrouped: true }],
-            loadingTimeout: undefined,
-            sorting: {
-                mode: 'none'
-            },
-            allowGrouping: true,
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2', field3: '34' }, { field1: '1', field2: '4', field3: '8' }]
-            }
-        }).dxDataGrid('instance');
-
-        $(dataGrid.$element())
-            .find('.dx-group-row')
-            .eq(1)
-            .find('td')
-            .first()
-            .trigger('dxcontextmenu');
-
-        $('.dx-datagrid .dx-menu-item') // click on "Ungroup this"
-            .eq(0)
-            .trigger('dxclick');
-
-        this.clock.tick(300);
-
-        assert.equal(dataGrid.columnOption('field2', 'groupIndex'), undefined, 'field2 has no groupIndex');
-        assert.equal(dataGrid.columnOption('field3', 'groupIndex'), 0, 'field3 save its groupIndex');
-    });
-
-    QUnit.test('Ungroup all columns via group row context menu', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', { dataField: 'field2', groupIndex: 1, showWhenGrouped: true }, { dataField: 'field3', groupIndex: 0, showWhenGrouped: true }],
-            loadingTimeout: undefined,
-            allowGrouping: true,
-            grouping: {
-                contextMenuEnabled: true
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2', field3: '34' }, { field1: '1', field2: '4', field3: '8' }]
-            }
-        }).dxDataGrid('instance');
-
-        $(dataGrid.$element())
-            .find('.dx-group-row td')
-            .eq(1)
-            .trigger('dxcontextmenu');
-
-        $('.dx-datagrid .dx-menu-item') // click on "clear groupings"
-            .eq(1)
-            .trigger('dxclick');
-
-        this.clock.tick(300);
-
-        assert.equal(dataGrid.columnOption('field2', 'groupIndex'), undefined, 'field2 has no groupIndex');
-        assert.equal(dataGrid.columnOption('field3', 'groupIndex'), undefined, 'field3 has no groupIndex');
-    });
-
-    QUnit.test('Grouping with context menu - check custom texts', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2', { dataField: 'field3', showWhenGrouped: true }],
-            loadingTimeout: undefined,
-            grouping: {
-                contextMenuEnabled: true,
-                texts: {
-                    groupByThisColumn: 'test1',
-                    ungroup: 'test2',
-                    ungroupAll: 'test3'
-                }
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2', field3: '34' }, { field1: '1', field2: '4', field3: '8' }]
-            }
-        }).dxDataGrid('instance');
-
-        $(dataGrid.$element())
-            .find('.dx-header-row td')
-            .eq(2)
-            .trigger('dxcontextmenu');
-
-        const $menuItems = $('.dx-datagrid .dx-menu-item');
-
-        assert.equal($menuItems.eq(3).text(), 'test1', 'Custom \'group\' message');
-        assert.equal($menuItems.eq(4).text(), 'test2', 'Custom \'ungroup\' message');
-        assert.equal($menuItems.eq(5).text(), 'test3', 'Custom \'cleanGrouping\' message');
-    });
-
-    QUnit.test('Context menu does not have grouping items when \'contextMenuEnabled\' is false', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2', 'field3'],
-            loadingTimeout: undefined,
-            allowGrouping: false,
-            grouping: {
-                contextMenuEnabled: false
-            },
-            dataSource: {
-                store: [{ field1: '1', field2: '2', field3: '34' }, { field1: '1', field2: '4', field3: '8' }]
-            }
-        }).dxDataGrid('instance');
-
-        $(dataGrid.$element())
-            .find('.dx-header-row td')
-            .eq(2)
-            .trigger('dxcontextmenu');
-
-        const $menuItems = $('.dx-datagrid .dx-menu-item');
-
-        assert.ok(!$menuItems.find(':contains(Group this)').length, 'Menu items doesn\'t contain \'group\' command');
-        assert.ok(!$menuItems.find(':contains(Ungroup this)').length, 'Menu items doesn\'t contain \'ungroup\' command');
-        assert.ok(!$menuItems.find(':contains(Clear grouping)').length, 'Menu items doesn\'t contain \'clear grouping\' command');
     });
 
     // T860356
@@ -1207,15 +696,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal(onOptionChangedCallCount, 1, 'onOptionChanged call count');
     });
 
-    function isColumnHidden($container, index) {
-        const $colsHeadersView = $container.find('.dx-datagrid-headers col');
-        const $colsRowsView = $container.find('.dx-datagrid-headers col');
-        const headersColWidth = $colsHeadersView.get(index).style.width;
-        const rowsViewWidth = $colsRowsView.get(index).style.width;
-
-        return (headersColWidth === '0.0001px' || headersColWidth === '0px') && (rowsViewWidth === '0.0001px' || rowsViewWidth === '0px');
-    }
-
     QUnit.test('Columns hiding - columnHidingEnabled is true', function(assert) {
     // arrange
         $('#container').width(200);
@@ -1235,9 +715,9 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
         // act
         assert.equal($visibleColumns.length, 3, 'only 1 column is visible');
-        assert.ok(!isColumnHidden(dataGrid, 0), 'first column is shown');
-        assert.ok(isColumnHidden(dataGrid, 1), 'second column is hidden');
-        assert.ok(!isColumnHidden(dataGrid, 2), 'adaptive column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(1), 'second column is hidden');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is shown');
         assert.equal($visibleColumns.eq(0).text(), 'First Name', 'it is \'firstName\' column');
         assert.equal(adaptiveColumnsController.getHiddenColumns()[0].dataField, 'lastName', '\'lastName\' column is hidden');
 
@@ -1249,9 +729,9 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
         // assert
         assert.equal($visibleColumns.length, 3, '2 columns are visible');
-        assert.ok(!isColumnHidden(dataGrid, 0), 'first column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 1), 'second column is shown');
-        assert.ok(isColumnHidden(dataGrid, 2), 'adaptive column is hidden');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is hidden');
         assert.equal($visibleColumns.eq(0).text(), 'First Name', 'First is \'firstName\' column');
         assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'Second is \'lastName\' column');
         assert.equal(adaptiveColumnsController.getHiddenColumns().length, 0, 'There is no hidden columns');
@@ -1276,9 +756,9 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         const $hiddenColumn = $('.dx-datagrid-hidden-column').eq(0);
 
         // act
-        assert.ok(isColumnHidden(dataGrid, 0), 'first column is hidden');
-        assert.ok(!isColumnHidden(dataGrid, 1), 'second column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 2), 'adaptive column is shown');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(0), 'first column is hidden');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is shown');
         assert.equal($visibleColumns.length, 3, 'only 1 column is visible');
         assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'it is \'lastName\' column');
         assert.equal(adaptiveColumnsController.getHiddenColumns()[0].dataField, 'firstName', '\'firstName\' column is hidden');
@@ -1294,52 +774,14 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         $visibleColumns = $(instance.$element().find('.dx-header-row td'));
 
         // assert
-        assert.ok(!isColumnHidden(dataGrid, 0), 'first column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 1), 'second column is shown');
-        assert.ok(isColumnHidden(dataGrid, 2), 'adaptive column is hidden');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is hidden');
         assert.equal($visibleColumns.length, 3, '2 columns are visible');
         assert.equal($visibleColumns.eq(0).text(), 'First Name', 'First is \'firstName\' column');
         assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'Second is \'lastName\' column');
         assert.equal(adaptiveColumnsController.getHiddenColumns().length, 0, 'There is no hidden columns');
         assert.equal(adaptiveColumnsController.getHidingColumnsQueue().length, 2, 'There is 2 columns in hiding queue');
-    });
-
-    QUnit.test('Columns hiding - grouping with hidingPriority', function(assert) {
-    // arrange
-        $('#container').width(600);
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ firstName: 'Blablabla', lastName: 'Psy', age: 40 }],
-            columns: [{ dataField: 'firstName', hidingPriority: 0, groupIndex: 0 }, { dataField: 'lastName', hidingPriority: 1 }, { dataField: 'age', hidingPriority: 2 }]
-        });
-        const instance = dataGrid.dxDataGrid('instance');
-        const adaptiveColumnsController = instance.getController('adaptiveColumns');
-        let $visibleColumns;
-
-        this.clock.tick();
-
-        $visibleColumns = $(instance.$element().find('.dx-header-row td:not(.dx-datagrid-group-space)'));
-
-        // act
-        assert.equal($visibleColumns.length, 3, '2 column are visible');
-        assert.equal($visibleColumns.eq(0).text(), 'Last Name', 'first is \'lastName\' column');
-        assert.equal($visibleColumns.eq(1).text(), 'Age', 'second is \'lastName\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns().length, 0, 'There no hidden columns');
-
-        $('#container').width(150);
-        instance.updateDimensions();
-        this.clock.tick();
-        $visibleColumns = $(instance.$element().find('.dx-header-row td:not(.dx-datagrid-group-space)'));
-
-        // assert
-        assert.ok(!isColumnHidden(dataGrid, 0), 'first column is shown');
-        assert.ok(isColumnHidden(dataGrid, 1), 'second column is hidden');
-        assert.ok(!isColumnHidden(dataGrid, 2), 'adaptive column is shown');
-        assert.equal($visibleColumns.length, 3, '1 column is visible');
-        assert.equal($visibleColumns.eq(1).text(), 'Age', 'it is \'age\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns().length, 1, 'There is 1 hidden columns');
-        assert.equal(adaptiveColumnsController.getHiddenColumns()[0].dataField, 'lastName', '\'lastName\' column is hidden');
     });
 
     QUnit.test('Columns hiding - column without priority must stay (hidingPriority)', function(assert) {
@@ -1359,11 +801,11 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         $visibleColumns = $(instance.$element().find('.dx-header-row td'));
 
         // act
-        assert.ok(isColumnHidden(dataGrid, 0), 'first column is hidden');
-        assert.ok(isColumnHidden(dataGrid, 1), 'second column is hidden');
-        assert.ok(!isColumnHidden(dataGrid, 2), 'third column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 3), 'fourth column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 4), 'adaptive column is shown');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(0), 'first column is hidden');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(1), 'second column is hidden');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'third column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(3), 'fourth column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(4), 'adaptive column is shown');
         assert.equal($visibleColumns.length, 5, 'only 2 columns are visible');
         assert.equal($visibleColumns.eq(2).text(), 'Age', 'First is \'age\' column');
         assert.equal($visibleColumns.eq(3).text(), 'Country', 'Second is \'country\' column');
@@ -1379,11 +821,11 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         $visibleColumns = $(instance.$element().find('.dx-header-row td'));
 
         // assert
-        assert.ok(!isColumnHidden(dataGrid, 0), 'first column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 1), 'second column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 2), 'third column is shown');
-        assert.ok(!isColumnHidden(dataGrid, 3), 'fourth column is shown');
-        assert.ok(isColumnHidden(dataGrid, 4), 'adaptive column is hidden');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'third column is shown');
+        assert.ok(!dataGridWrapper.headers.isColumnHidden(3), 'fourth column is shown');
+        assert.ok(dataGridWrapper.headers.isColumnHidden(4), 'adaptive column is hidden');
         assert.equal($visibleColumns.length, 5, '4 columns are visible');
         assert.equal($visibleColumns.eq(0).text(), 'First Name', 'First is \'firstName\' column');
         assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'Second is \'lastName\' column');
@@ -2053,34 +1495,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         // assert
         assert.ok($dataGrid.find('.dx-datagrid-rowsview').height(), 'rowsView has height');
         $('#qunit-fixture').removeClass('qunit-fixture-auto-height');
-    });
-
-    // T706583
-    QUnit.test('grouping if columnRenderingMode is virtual, filterRow is visible and datetime column exists', function(assert) {
-    // arrange, act
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: [{}],
-            loadingTimeout: undefined,
-            scrolling: {
-                columnRenderingMode: 'virtual'
-            },
-            columnWidth: 100,
-            width: 500,
-            filterRow: {
-                visible: true
-            },
-            columns: [{
-                dataField: 'c1',
-                dataType: 'datetime'
-            },
-            'c2', 'c3', 'c4', 'c5', 'c6']
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.columnOption('c2', 'groupIndex', 0);
-
-        // assert
-        assert.equal(dataGrid.getVisibleColumns()[0].type, 'groupExpand', 'grouping is applied');
     });
 
     QUnit.test('DataGrid should apply columns that are dynamically added to a band (T815945)', function(assert) {
@@ -2995,69 +2409,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal($errorRow.find('.dx-error-message > a').attr('href'), errorUrl, 'Url error code');
     });
 
-    // T817255
-    QUnit.test('No error after ungrouping with custom store and column reordering', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            columns: ['field1', {
-                dataField: 'field2',
-                groupIndex: 0
-            }],
-            groupPanel: { visible: true },
-            allowColumnReordering: true,
-            dataSource: {
-                key: 'field1',
-                load: function() {
-                    return [{ field1: 1, field2: 1 }, { field1: 2, field2: 2 }];
-                }
-            }
-        });
-
-        this.clock.tick();
-
-        const columnController = dataGrid.getController('columns');
-
-        // act
-        columnController.moveColumn(0, 1, 'group', 'headers');
-
-        // assert
-        assert.strictEqual($($(dataGrid.$element()).find('.dx-error-row')).length, 0, 'no errors');
-    });
-
-    // T819729
-    QUnit.test('correct cellInfo is passed to cellTemplate function after ungrouping', function(assert) {
-    // arrange
-        let cellTemplateCallCount = 0;
-        const dataGrid = createDataGrid({
-            dataSource: [{ field1: 'some', field2: 'some' }, { field1: 'some', field2: 'some' }],
-            allowColumnReordering: true,
-            columns: [{
-                dataField: 'field1',
-                groupIndex: 0,
-                cellTemplate: function(cellElement, cellInfo) {
-                // assert
-                    cellTemplateCallCount++;
-                    assert.notOk(cellInfo.data.key);
-                    assert.notOk(cellInfo.data.items);
-                    assert.ok(cellInfo.data.field1);
-                    assert.ok(cellInfo.data.field2);
-                }
-            }, 'field2']
-        });
-
-        this.clock.tick();
-
-        const columnController = dataGrid.getController('columns');
-
-        // act
-        columnController.moveColumn(0, 1, 'group', 'headers');
-        cellTemplateCallCount = 0;
-        this.clock.tick();
-
-        // assert
-        assert.equal(cellTemplateCallCount, 2, 'cellTemplate call count');
-    });
-
     QUnit.test('Not raise error if key field is null', function(assert) {
     // act
         const dataGrid = createDataGrid({
@@ -3071,33 +2422,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         // assert
         const $errorRow = $($(dataGrid.$element()).find('.dx-error-row'));
         assert.equal($errorRow.length, 0, 'error row is not shown');
-    });
-
-    // T481276
-    QUnit.test('updateDimensions during grouping when fixed to right column exists', function(assert) {
-        let loadResult = $.Deferred().resolve([{}]);
-        const dataGrid = createDataGrid({
-            columns: ['field1', 'field2', { dataField: 'field3', fixed: true, fixedPosition: 'right' }],
-            loadingTimeout: undefined,
-            cacheEnabled: false,
-            dataSource: {
-                load: function() {
-                    return loadResult;
-                }
-            }
-        });
-
-
-        loadResult = $.Deferred();
-        dataGrid.columnOption('field1', 'groupIndex', 0);
-
-        // act
-        dataGrid.updateDimensions();
-        loadResult.resolve([{}]);
-
-        // assert
-        assert.ok(dataGrid.isReady(), 'dataGrid is ready');
-        assert.strictEqual($(dataGrid.$element()).find('.dx-group-row').length, 2, 'grouped rows are rendered');
     });
 
     // T334530
@@ -3722,36 +3046,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.ok($(dataGrid.$element()).hasClass('dx-state-disabled'), 'dx-state-disabled class added');
     });
 
-    // T697860
-    QUnit.test('dataSource change with grouping and columns should force one loading only', function(assert) {
-    // arrange, act
-        const loadingSpy = sinon.spy();
-
-        const options = {
-            dataSource: new DataSource([{ field1: 1 }]),
-            columns: ['field1']
-        };
-
-        const dataGrid = createDataGrid(options);
-
-        this.clock.tick(0);
-
-        options.dataSource.store().on('loading', loadingSpy);
-
-        // act
-        options.dataSource = new DataSource([{ field1: 2 }]);
-        options.dataSource.store().on('loading', loadingSpy);
-        options.grouping = {};
-        options.paging = {};
-
-        dataGrid.option(options);
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(loadingSpy.callCount, 1, 'loading called once');
-        assert.deepEqual(dataGrid.getVisibleRows()[0].data, { field1: 2 }, 'data is updated');
-    });
-
     QUnit.test('dataSource pageSize change', function(assert) {
     // arrange, act
         const dataGrid = createDataGrid({
@@ -3856,30 +3150,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.strictEqual($headerElements.first().text(), 'Column 3', 'text of the first header');
     });
 
-    // TODO this test without clock
-    // T197089
-    QUnit.test('group command column width after grouping column with showWhenGrouped', function(assert) {
-    // arrange
-    // act
-        const $dataGrid = $('#dataGridWithStyle').dxDataGrid({
-            dataSource: [{ field1: '1' }],
-            columnAutoWidth: true,
-            columns: [{ dataField: 'field1', showWhenGrouped: true }]
-        });
-
-        this.clock.tick();
-
-        // act
-        $dataGrid.dxDataGrid('instance').columnOption('field1', 'groupIndex', 0);
-
-        this.clock.tick();
-
-        // assert
-        const cols = $dataGrid.find('colgroup').first().children();
-
-        assert.ok(Math.abs(30 - cols.eq(0).width()) <= 1, 'grouped column width');
-    });
-
     // T196532
     QUnit.test('columns change when changed dataSource parameters', function(assert) {
     // arrange, act
@@ -3904,22 +3174,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         // assert
         assert.equal(loadingCount, 2, 'second load for apply sorting');
         assert.equal(dataGrid.getController('data').items()[0].data.b, 1);
-    });
-
-    // T708525
-    QUnit.test('change columns with grouping after dataSource change', function(assert) {
-        // arrange, act
-        const dataGrid = createDataGrid({});
-
-        // act
-        dataGrid.option('dataSource', [{ a: 1, b: 2 }]);
-        dataGrid.option('columns', ['a', { dataField: 'b', groupIndex: 0 }]);
-
-        this.clock.tick();
-
-        // assert
-        assert.equal(dataGrid.getVisibleRows()[0].rowType, 'group', 'first row type is group');
-        assert.equal(dataGrid.columnOption('b', 'groupIndex'), 0, 'column b is grouped');
     });
 
     QUnit.test('Column changes are applied while dataSource is loading (T895552)', function(assert) {
@@ -4534,57 +3788,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.equal(dataGrid._views.columnHeadersView.element().find('td').length, 1, 'count columns');
     });
 
-    QUnit.test('Correct update group panel items runtime', function(assert) {
-    // arrange
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            columns: ['field1', 'field2'],
-            groupPanel: { visible: true },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '3', field2: '4' }, { field1: '5', field2: '6' }]
-            }
-        }).dxDataGrid('instance');
-
-        this.clock.tick();
-
-        // act
-        dataGrid.columnOption(0, { groupIndex: 0 });
-        this.clock.tick();
-
-        // assert
-        const $groupPanelItems = $('#dataGrid').find('.dx-group-panel-item');
-
-        assert.equal($groupPanelItems.length, 1, 'count of group panel items');
-    });
-
-    QUnit.test('Check group panel items are draggable when toolbar items updated runtime', function(assert) {
-    // arrange
-        let renderCompletedCallCount = 0;
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            onInitialized: function(e) {
-                e.component.getView('headerPanel').renderCompleted.add(function() {
-                    renderCompletedCallCount++;
-                });
-            },
-            columns: [{ dataField: 'field1', groupIndex: 0 }, 'field2'],
-            groupPanel: { visible: true },
-            dataSource: {
-                store: [{ field1: '1', field2: '2' }, { field1: '3', field2: '4' }, { field1: '5', field2: '6' }]
-            }
-        }).dxDataGrid('instance');
-
-        this.clock.tick();
-
-        // act
-        const toolbar = dataGrid.$element().find('.dx-toolbar').dxToolbar('instance');
-        toolbar.option('items', toolbar.option('items'));
-        this.clock.tick();
-
-        // assert
-        const $groupPanelItems = $('#dataGrid').find('.dx-toolbar .dx-datagrid-drag-action');
-        assert.equal($groupPanelItems.length, 1, 'count of group panel items');
-        assert.equal(renderCompletedCallCount, 3, 'renderCompleted call count');
-    });
-
     // T113684
     QUnit.test('Height rows view = height content', function(assert) {
     // arrange, act
@@ -4768,38 +3971,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.equal($dataGrid.dxDataGrid('instance').getView('rowsView').getTableElements().length, 1, 'one main rowsview table');
     });
 
-    // T445971
-    QUnit.test('Hide group panel and search panel when calculateDisplayValue is defined', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [
-                { field1: 1, field2: 'test1', field3: 'test2' },
-                { field1: 2, field2: 'test3', field3: 'test4' }
-            ],
-            columns: [{ dataField: 'field1', calculateDisplayValue: commonUtils.noop, groupIndex: 0 }, { dataField: 'field2', groupIndex: 1 }],
-            groupPanel: { visible: true }
-        });
-
-        // act
-        dataGrid.option({
-            groupPanel: {
-                visible: false
-            },
-            searchPanel: {
-                visible: false
-            }
-        });
-
-        // assert
-        const visibleColumns = dataGrid.getVisibleColumns();
-        assert.strictEqual(visibleColumns.length, 3, 'count column');
-        assert.strictEqual(visibleColumns[0].dataField, 'field1', 'dataField of the first column');
-        assert.strictEqual(visibleColumns[0].groupIndex, 0, 'groupIndex of the first column');
-        assert.strictEqual(visibleColumns[1].dataField, 'field2', 'dataField of the second column');
-        assert.strictEqual(visibleColumns[1].groupIndex, 1, 'groupIndex of the second column');
-    });
-
     // T689294
     QUnit.test('onContentReady when there is no dataSource and stateStoring is enabled', function(assert) {
     // arrange
@@ -4944,25 +4115,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal($('#dataGrid').find('td').eq(1).find('.dx-datagrid-text-content').first().text(), 'Test Column');
     });
 
-    QUnit.test('Group row has correct text-align in RTL', function(assert) {
-    // arrange, act
-
-        const dataGrid = createDataGrid({
-            rtlEnabled: true,
-            loadingTimeout: undefined,
-            dataSource: {
-                store: [{ field1: '1', field2: '2', field3: '3', field4: '4', field5: '5' }]
-            },
-            columns: ['field1', 'field2', { dataField: 'field3', groupIndex: 0 }]
-        });
-        const groupedRows = $(dataGrid.$element()).find('.dx-group-row');
-        const cells = groupedRows.children();
-
-        // assert
-        assert.ok(groupedRows.length, 'We have grouped row');
-        assert.equal(cells.eq(1).css('textAlign'), 'right', 'Grouped cell has correct text-align');
-    });
-
     QUnit.test('expandAll', function(assert) {
     // arrange, act
         let expandAllGroupIndex;
@@ -4999,54 +4151,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
 
         // assert
         assert.equal(collapseAllGroupIndex, 1);
-    });
-
-    // T152353
-    QUnit.test('Stretch column datagrid width when grouping columns', function(assert) {
-    // arrange
-        const $dataGrid = $('#dataGridWithStyle').dxDataGrid({
-            dataSource: [{ field1: '1', field2: '2', field3: '3', field4: '4', field5: '5' }],
-            groupPanel: {
-                visible: true
-            },
-            columns: [
-                {
-                    dataField: 'field1',
-                    width: 130
-                },
-                {
-                    dataField: 'field2'
-                },
-                {
-                    dataField: 'field3',
-                    groupIndex: 0
-                }
-            ],
-            allowColumnReordering: true,
-            allowColumnResizing: true
-        });
-        const dataGrid = $dataGrid.dxDataGrid('instance');
-        const columnController = dataGrid.getController('columns');
-
-        this.clock.tick();
-        // act
-        const gridInitialWidth = $dataGrid.outerWidth();
-
-        columnController.moveColumn(2, 0, 'headers', 'group');
-        this.clock.tick();
-
-        const gridWidthAfterGrouping = $dataGrid.outerWidth();
-
-        columnController.moveColumn(0, 1, 'group', 'headers');
-        this.clock.tick();
-
-        const gridWidthAfterUngrouping = $dataGrid.outerWidth();
-
-        // assert
-        // TODO: if we set style or rule to grid's container, this asserts will be "equal" instead of "ok"
-        assert.equal(gridWidthAfterGrouping, gridInitialWidth, 'After grouping columns grid width equals initial grid width');
-        assert.equal(gridWidthAfterUngrouping, gridWidthAfterGrouping, 'After move one group column to grid, grid size equals than previous grouping');
-        assert.equal(gridWidthAfterUngrouping, gridInitialWidth);
     });
 
     // B239291
@@ -5608,69 +4712,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         // assert
         const $firstCell = $($(dataGrid.$element()).find('.dx-data-row').eq(0).children().eq(0));
         assert.equal($firstCell.text(), 'Test 1', 'Lookup text is correct');
-    });
-
-    // T800495
-    QUnit.test('The calculateCellValue arguments should be correct after resetting the state when there is a grouped column', function(assert) {
-    // arrange
-        const calculateCellValue = sinon.spy();
-        const dataGrid = createDataGrid({
-            columns: [{ dataField: 'field1', groupIndex: 0, calculateCellValue: calculateCellValue }, 'field2'],
-            dataSource: [{ field1: 'test1', field2: 'test2' }, { field1: 'test3', field2: 'test4' }]
-        });
-
-        this.clock.tick(0);
-        calculateCellValue.reset();
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        // assert
-        assert.deepEqual(calculateCellValue.getCall(0).args[0], { field1: 'test1', field2: 'test2' }, 'calculateCellValue - first call arguments');
-    });
-
-    // T817555
-    QUnit.test('State reset should save default grouping', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            columns: [{ dataField: 'field1', groupIndex: 0 }, 'field2'],
-            dataSource: [{ field1: 'test1', field2: 'test2' }, { field1: 'test3', field2: 'test4' }]
-        });
-
-        this.clock.tick(0);
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(dataGrid.columnOption(0, 'groupIndex'), 0, 'groupIndex was not reset');
-    });
-
-    // T817555
-    QUnit.test('Double reset should work correctly when rows are grouped', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            columns: [{ dataField: 'field1', groupIndex: 0 }, { dataField: 'field2', sortOrder: 'asc' }],
-            dataSource: [{ field1: 'test1', field2: 'test2' }, { field1: 'test3', field2: 'test4' }]
-        });
-
-        this.clock.tick(0);
-
-        // act
-        dataGrid.columnOption(0, 'groupIndex', undefined);
-        dataGrid.columnOption(1, 'sortOrder', undefined);
-
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(dataGrid.columnOption(0, 'groupIndex'), 0, 'groupIndex was returned to default');
-        assert.equal(dataGrid.columnOption(1, 'sortOrder'), 'asc', 'sortOrder was returned to default');
     });
 
     QUnit.test('Clear state when initial options is defined in dataSource', function(assert) {
@@ -6306,47 +5347,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.strictEqual(dataGrid.getVisibleRows()[1].data.detail, 'updated', 'row 1 data is updated');
     });
 
-    // T720597
-    QUnit.test('Grouping and ungrouping', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [
-                { id: 1, col1: '1 1', col2: '1 2', col3: '1 3' },
-                { id: 2, col1: '2 1', col2: '2 2', col3: '2 3' },
-                { id: 3, col1: '3 1', col2: '3 2', col3: '3 3' }
-            ],
-            loadingTimeout: undefined,
-            paging: {
-                pageSize: 2
-            },
-            repaintChangesOnly: true,
-            scrolling: {
-                mode: 'virtual'
-            },
-            columns: [
-                { dataField: 'col1', showWhenGrouped: true },
-                { dataField: 'col2', showWhenGrouped: true },
-                { dataField: 'col3', showWhenGrouped: true }
-            ],
-            summary: {
-                groupItems: [{
-                    column: 'Col1',
-                    summaryType: 'count'
-                }]
-            }
-        });
-
-        // act
-        dataGrid.columnOption('col1', 'groupIndex', 0);
-        dataGrid.columnOption('col2', 'groupIndex', 1);
-        dataGrid.columnOption('col1', 'groupIndex', undefined);
-
-        // assert
-        assert.strictEqual($(dataGrid.element()).find('.dx-datagrid-headers td').length, 4, 'header cell count is correct');
-        assert.strictEqual($(dataGrid.getRowElement(0)).children().length, 2, 'data cell count for first group row is correct');
-        assert.strictEqual($(dataGrid.getRowElement(1)).children().length, 4, 'data cell count for second data row is correct');
-    });
-
     QUnit.test('watch in cellPrepared should works after push', function(assert) {
     // arrange
         const dataGrid = createDataGrid({
@@ -6430,33 +5430,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(cellPreparedArgs[0].columnIndex, 1, 'cell prepared columnIndex');
         assert.equal(cellPreparedArgs[0].value, 'updated', 'cell prepared value');
         assert.equal(cellPreparedArgs[0].oldValue, 'test1', 'cell prepared oldValue');
-    });
-
-    QUnit.test('Column widths should be updated after expand group row if repaintChangesOnly is true', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            keyExpr: 'id',
-            dataSource: [
-                { id: 1, group: 'group1' },
-                { id: 2, group: 'group1' },
-                { id: 3, group: 'group2' },
-                { id: 4, group: 'group2' }
-            ],
-            grouping: {
-                autoExpandAll: false
-            },
-            columns: ['id', {
-                dataField: 'group',
-                groupIndex: 0
-            }],
-            repaintChangesOnly: true
-        });
-
-        dataGrid.expandRow(['group1']);
-
-        // assert
-        assert.equal(dataGrid.getVisibleColumns()[0].visibleWidth, 30, 'visibleWidth for first groupExpand column');
     });
 
     QUnit.test('push changes for adaptive row', function(assert) {
@@ -6653,51 +5626,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.strictEqual(columns[0].width, 200, 'width of the first column');
         assert.strictEqual(columns[1].width, 200, 'width of the second column');
         assert.strictEqual(columns[2].width, 200, 'width of the third column');
-    });
-
-    QUnit.test('Group Row - expandRow should resolve its promise only after re-rendering (T880769)', function(assert) {
-        // arrange
-        const getRowsInfo = function(element) {
-            const $rows = $(element).find('.dx-datagrid-rowsview .dx-row[role=\'row\']');
-            return {
-                rowCount: $rows.length,
-                groupRow: $($rows.eq(0)).hasClass('dx-group-row'),
-                dataRow: $($rows.eq(1)).hasClass('dx-data-row')
-            };
-        };
-        const dataGrid = createDataGrid({
-            dataSource: [
-                { column1: 'value1', column2: 'value2' }
-            ],
-            grouping: {
-                autoExpandAll: false,
-            },
-            columns: [
-                {
-                    dataField: 'column1',
-                    groupIndex: 0
-                },
-                'column2'
-            ],
-            onRowExpanded: function() {
-                const info = getRowsInfo(dataGrid.element());
-                assert.step(`rowExpanded rowCount: ${info.rowCount}, groupRow: ${info.groupRow}, dataRow: ${info.dataRow}`);
-            }
-        });
-        this.clock.tick();
-
-        // act
-        dataGrid.expandRow(['value1']).done(() => {
-            const info = getRowsInfo(dataGrid.element());
-
-            assert.step(`done rowCount: ${info.rowCount}, groupRow: ${info.groupRow}, dataRow: ${info.dataRow}`);
-        });
-        this.clock.tick();
-
-        assert.verifySteps([
-            'rowExpanded rowCount: 2, groupRow: true, dataRow: true',
-            'done rowCount: 2, groupRow: true, dataRow: true'
-        ]);
     });
 });
 
@@ -6981,59 +5909,6 @@ QUnit.module('columnWidth auto option', {
         const secondColumnWidth = $($('#dataGrid').find('.dx-datagrid-headers').find('td')[1]).width();
 
         assert.ok(secondColumnWidth > 2 * firstColumnWidth, 'second column width more then first');
-    });
-
-    QUnit.test('Group cell should not have width style', function(assert) {
-        const dataSource = [
-            { firstName: 'Alex', lastName: 'Black', room: 903 },
-            { firstName: 'Alex', lastName: 'White', room: 904 }
-        ];
-
-        // act
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: dataSource,
-            columnAutoWidth: true,
-            columns: [{
-                dataField: 'firstName',
-                groupIndex: 0,
-                width: 100,
-            }, {
-                dataField: 'lastName',
-                width: 150
-            }, 'room']
-        }).dxDataGrid('instance');
-
-        // assert
-        assert.equal($(dataGrid.getCellElement(0, 1)).get(0).style.width, '', 'width style is not defined for group cell');
-        assert.equal($(dataGrid.getCellElement(1, 1)).get(0).style.width, '150px', 'width style is defined for data cell');
-    });
-
-    // T661361
-    QUnit.test('Group space cells should have correct width if data rows are not visible', function(assert) {
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ group1: 1, group2: 1, id: 1 }],
-            grouping: {
-                autoExpandAll: false
-            },
-            columns: ['id', {
-                dataField: 'group1',
-                groupIndex: 0
-            }, {
-                dataField: 'group2',
-                groupIndex: 1
-            }]
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.expandRow([1]);
-        dataGrid.updateDimensions();
-
-        // assert
-        const $groupSpaceCells = $(dataGrid.getRowElement(1)).children('.dx-datagrid-group-space');
-        assert.equal($groupSpaceCells.length, 2, 'two group space cells in second row');
-        assert.equal($groupSpaceCells.eq(0).width(), $groupSpaceCells.eq(1).width(), 'group space cell widths are equals');
     });
 
     QUnit.test('Check table params with set width', function(assert) {
@@ -7647,33 +6522,6 @@ QUnit.module('Row dragging', baseModuleConfig, () => {
                 }
             });
         }
-    });
-
-    QUnit.test('GroupPanel items should be visible by adding new one (T880880)', function(assert) {
-        // arrange
-        const headerPanelWrapper = dataGridWrapper.headerPanel;
-        createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ field1: 1, field2: 2, field3: 3 }],
-            groupPanel: {
-                visible: true
-            },
-            columns: [
-                { dataField: 'field1', groupIndex: 0 },
-                { dataField: 'field2' }
-            ]
-        });
-
-        // act
-        const groupPanel = headerPanelWrapper.getGroupPanelElement();
-        $('<div>').addClass('dx-group-panel-item').text('test').appendTo(groupPanel);
-
-        const items = headerPanelWrapper.getGroupPanelItems();
-        const itemsWidth = items.eq(0).outerWidth(true) + items.eq(1).outerWidth(true);
-
-        // assert
-        assert.equal(items.length, 2, '2 items in group panel');
-        assert.roughEqual(groupPanel.innerWidth(), itemsWidth, 1.01, 'enough space for children display');
     });
 });
